fix(billing): extract makeOnStepFinishHandler + fix LanguageModelV2Usage field names
Next.js route files may only export HTTP verbs + a small set of config
fields; exporting makeOnStepFinishHandler from route.ts caused the build
to fail with "not a valid Route export field". Fix: move the function to
a sibling step-finish-handler.ts and import it from route.ts.

Also fix the usage field names: AI SDK v5 (LanguageModelV2Usage) uses
inputTokens/outputTokens, not promptTokens/completionTokens. Updated
calcStepCostDollars signature and all callers/tests accordingly.
